Create ImapConfig for IMAP-related debugging options
