+++ k9mail-library/src/main/java/com/fsck/k9/mail/K9MailLib.java
@@ -3,7 +3,6 @@ package com.fsck.k9.mail;
 
 public class K9MailLib {
     private static DebugStatus debugStatus = new DefaultDebugStatus();
-    private static ImapExtensionStatus imapExtensionStatus = new DefaultImapExtensionStatus();
 
     private K9MailLib() {
     }
@@ -43,14 +42,6 @@ public class K9MailLib {
         return debugStatus.debugSensitive();
     }
 
-    public static boolean shouldUseCondstore() {
-        return imapExtensionStatus.useCondstore();
-    }
-
-    public static boolean shouldUseQresync() {
-        return imapExtensionStatus.useQresync();
-    }
-
     public static void setDebugSensitive(boolean b) {
         if (debugStatus instanceof WritableDebugStatus) {
             ((WritableDebugStatus) debugStatus).setSensitive(b);
@@ -63,28 +54,10 @@ public class K9MailLib {
         }
     }
 
-    public static void setUseCondstore(boolean useCondstore) {
-        if (imapExtensionStatus instanceof WritableImapExtensionStatus) {
-            ((WritableImapExtensionStatus) imapExtensionStatus).setUseCondstore(useCondstore);
-        }
-    }
-
-    public static void setUseQresync(boolean useQresync) {
-        if (imapExtensionStatus instanceof WritableImapExtensionStatus) {
-            ((WritableImapExtensionStatus) imapExtensionStatus).setUseQresync(useQresync);
-        }
-    }
-
     public interface DebugStatus {
         boolean enabled();
 
         boolean debugSensitive();
-    }
-
-    public interface ImapExtensionStatus {
-        boolean useCondstore();
-
-        boolean useQresync();
     }
 
     public static void setDebugStatus(DebugStatus status) {
@@ -94,23 +67,10 @@ public class K9MailLib {
         debugStatus = status;
     }
 
-    public static void setImapExtensionStatus(ImapExtensionStatus status) {
-        if (status == null) {
-            throw new IllegalArgumentException("status cannot be null");
-        }
-        imapExtensionStatus = status;
-    }
-
     private interface WritableDebugStatus extends DebugStatus {
         void setEnabled(boolean enabled);
 
         void setSensitive(boolean sensitive);
-    }
-
-    private interface WritableImapExtensionStatus extends ImapExtensionStatus {
-        void setUseCondstore(boolean useCondstore);
-
-        void setUseQresync(boolean useQresync);
     }
 
     private static class DefaultDebugStatus implements WritableDebugStatus {
@@ -135,31 +95,6 @@ public class K9MailLib {
         @Override
         public void setSensitive(boolean sensitive) {
             this.sensitive = sensitive;
-        }
-    }
-
-    private static class DefaultImapExtensionStatus implements WritableImapExtensionStatus {
-        private boolean useCondstore = true;
-        private boolean useQresync = true;
-
-        @Override
-        public boolean useCondstore() {
-            return useCondstore;
-        }
-
-        @Override
-        public boolean useQresync() {
-            return useQresync;
-        }
-
-        @Override
-        public void setUseCondstore(boolean useCondstore) {
-            this.useCondstore = useCondstore;
-        }
-
-        @Override
-        public void setUseQresync(boolean useQresync) {
-            this.useQresync = useQresync;
         }
     }
 }
