+++ k9mail-library/src/main/java/com/fsck/k9/mail/store/imap/SelectOrExamineResponse.java
@@ -48,7 +48,7 @@ class SelectOrExamineResponse {
         }
         this.readWriteMode = isModeReadWriteIfAvailable(ImapUtility.getLastResponse(imapResponses));
         if (folder.doesConnectionSupportQresync() && highestModSeq != INVALID_HIGHEST_MOD_SEQ
-                && K9MailLib.shouldUseQresync()) {
+                && ImapConfig.shouldUseQresync()) {
             this.qresyncParamResponse = QresyncParamResponse.fromSelectOrExamineResponse(imapResponses, folder);
         } else {
             this.qresyncParamResponse = null;
